+++ js/app.js
@@ -177,7 +177,7 @@ function App () {
 				moduleFolder = defaultModuleFolder;
 			} else {
 				// running in Jest, allow defaultModules placed under moduleDir for testing
-				if (env.modulesDir === "modules") {
+				if (env.modulesDir === "modules" || env.modulesDir === "tests/mocks") {
 					moduleFolder = defaultModuleFolder;
 				}
 			}