+++ CHANGELOG.md
@@ -25,8 +25,9 @@ planned for 2025-04-01
 
 ### Fixed
 
-- [calendar] Fix clipping events being broadcast (#3678)
+- [calendar] fix clipping events being broadcast (#3678)
 - [tests] Electron tests: Fixes for running under new github image ubuntu-24.04, replace xserver with labwc (#3676)
+- [calendar] fix arrayed symbols, #3267, again, add testcase, add testcase for #3678
 
 ## [2.30.0] - 2025-01-01